requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
mediapipe==0.10.21
numpy
tensorflow
scikit-learn==1.3.0
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
pillow==10.0.0
matplotlib==3.7.2 